+++ tests/OnRamp.Test/Config/EntityConfigEditor.cs
@@ -1,13 +1,15 @@
 ﻿using OnRamp.Config;
+using System.Threading.Tasks;
 
 namespace OnRamp.Test.Config
 {
     public class EntityConfigEditor : IConfigEditor
     {
-        public void BeforePrepare(IRootConfig config)
+        public Task BeforePrepareAsync(IRootConfig config)
         {
             var ec = config as EntityConfig;
             ec.Name = ec.Name.ToUpperInvariant();
+            return Task.CompletedTask;
         }
     }
 }